youtube-video-upload/requirements.txt
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-cloud-storage>=2.0.0
python-dotenv>=1.0.0
